Eliminate verus spec drift so CI catches real regressions
The sync checker reported 71 drifts at 34.8% coverage, which meant it
was effectively useless as a gate — too noisy to distinguish real
regressions from accumulated staleness. Most "drifts" were false
positives from the checker itself rather than genuine spec divergence.

Three categories of fixes bring coverage to 100% with 0 drifts:

Spec updates across aspen-coordination and aspen-raft verus files to
match canonical production code. Signature drifts from renamed types
(ElectionStateSpec → LeadershipState, LockEntrySpec → LockEntry),
changed parameters (f32 jitter factors replacing u32 percents), and
evolved function signatures. Body drifts from production code adding
debug_assert! guards, intermediate variable bindings, and reordered
control flow that verus specs hadn't tracked. Functions with f32 or
HashMap parameters that Verus can't reason about are now marked
external_body.

Missing production functions added to aspen-raft's src/verified/ for
39 exec functions that existed in verus specs but had never been ported
to the write_batcher, integrity, membership, and kv modules.

verus-metrics tooling improvements that caused most of the false
positives: recursive directory traversal so src/verified/queue/
subdirectory is no longer invisible, debug_assert!/assert! stripping
in the body normalizer, trailing let-binding normalization, std path
prefix and underscore-prefixed param normalization in type comparison,
name-collision resolution preferring signature-compatible matches, and
correct brace tracking in ensures clauses containing match expressions.

+++ crates/aspen-raft/verus/membership_spec.rs
@@ -225,6 +225,7 @@ verus! {
     /// # Returns
     ///
     /// Minimum cluster size, saturating at u32::MAX.
+    #[inline]
     pub fn min_cluster_size_for_tolerance(fault_tolerance: u32) -> (result: u32)
         ensures
             fault_tolerance as int * 2 + 1 <= u32::MAX as int ==>
@@ -232,16 +233,15 @@ verus! {
             fault_tolerance as int * 2 + 1 > u32::MAX as int ==>
                 result == u32::MAX
     {
-        // Manual saturating_mul(2) then saturating_add(1)
         let doubled = if fault_tolerance > u32::MAX / 2 {
             u32::MAX
         } else {
-            (fault_tolerance * 2) as u32
+            fault_tolerance * 2
         };
         if doubled > u32::MAX - 1 {
             u32::MAX
         } else {
-            (doubled + 1) as u32
+            doubled + 1
         }
     }
